Move internal function and class to header file
